DailySurveySummaryFrequencySegmentSelected moved to its own file
